Move get_binary_arches into common, and code sign every binary in the app bundle
